requirements.txt
anyio==4.12.1
async-timeout==4.0.3
beautifulsoup4==4.14.3
bs4==0.0.2
fastapi==0.135.0
langchain
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.15
langchain-huggingface==1.2.0
langchain-milvus==0.3.3
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
markdown-it-py==4.0.0
milvus-lite==2.5.1
openai==2.23.0
openai-agents==0.10.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pymilvus==2.6.9
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
sentence-transformers
torch --extra-index-url https://download.pytorch.org/whl/cpu